requirements.txt
tabula-py
PyPDF2
pandas
numpy
